Termino repository de productos

+++ src/main/java/com/ignaciosaccomano/proyecto_market/persistance/entity/Producto.java
@@ -9,7 +9,7 @@ public class Producto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_producto")
-    private Long idProducto;
+    private Integer idProducto;
 
     private String nombre;      // Mismo nombre que en la tabla, no es necesario el decorator @Column.
 
@@ -34,11 +34,11 @@ public class Producto {
     // Seteamos getters y setters
 
 
-    public Long getIdProducto() {
+    public Integer getIdProducto() {
         return idProducto;
     }
 
-    public void setIdProducto(Long idProducto) {
+    public void setIdProducto(Integer idProducto) {
         this.idProducto = idProducto;
     }
